+++ CHANGES.rst
@@ -5,6 +5,9 @@ History
 Next Release
 ============
 
+Python 2.6 is no longer supported.  We now test under cpython 2.7
+and cpython 3.3 through 3.6.
+
 Bugs Fixed
 ----------